+++ src/cdogs/bullet_class.h
@@ -72,7 +72,7 @@ typedef struct
 	int RangeLow;
 	int RangeHigh;
 	int Power;
-	double Mass;
+	float Mass;
 	Vec2i Size;
 	special_damage_e Special;
 	bool HurtAlways;
