+++ src/cdogs/objs.h
@@ -105,7 +105,7 @@ bool HasHitSound(
 void Damage(
 	const struct vec hitVector,
 	const int power,
-	const double mass,
+	const float mass,
 	const int flags, const int playerUID, const int uid,
 	const TileItemKind targetKind, const int targetUID,
 	const special_damage_e special);
